pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "flask-headless-auth"
version = "0.1.0"
description = "Modern, headless authentication for Flask APIs. JWT + OAuth + RBAC in one line."
readme = "README.md"
authors = [
    { name = "Dhruv Agnihotri", email = "dagni@umich.edu" }
]
license = "MIT"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Framework :: Flask",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Internet :: WWW/HTTP :: Session",
]
keywords = [
    "flask",
    "authentication",
    "jwt",
    "oauth",
    "api",
    "headless",
    "react",
    "rbac",
    "mfa",
    "nextjs",
    "spa"
]
dependencies = [
    "Flask>=2.0.0",
    "Flask-SQLAlchemy>=3.0.0",
    "Flask-JWT-Extended>=4.5.0",
    "Flask-Limiter>=3.3.0",
    "Flask-Caching>=2.0.0",
    "Flask-CORS>=4.0.0",
    "Flask-WTF>=1.1.0",
    "Flask-Talisman>=1.0.0",
    "Authlib>=1.2.0",
    "PyJWT>=2.8.0",
    "python-dotenv>=1.0.0",
    "PyYAML>=6.0",
    "Werkzeug>=2.0.0",
    "itsdangerous>=2.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.7.0",
    "flake8>=6.1.0",
]
email = [
    "brevo-python>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/Dhruvagnihotri/flask-headless-auth"
Repository = "https://github.com/Dhruvagnihotri/flask-headless-auth"
Issues = "https://github.com/Dhruvagnihotri/flask-headless-auth/issues"
Documentation = "https://github.com/Dhruvagnihotri/flask-headless-auth#readme"

[tool.setuptools.packages.find]
where = ["."]
include = ["flask_headless_auth*"]
exclude = ["tests*", "examples*", "instance*"]

[tool.setuptools.package-data]
flask_headless_auth = ["*.yml", "*.yaml"]
